+++ test/support/minitest_helpers.rb
@@ -74,7 +74,8 @@ module MinitestHelpers
 
     failed_patterns = []
     patterns_to_match.each do |pattern|
-      failed_patterns << pattern if queries.none? { |sql| sql.include?(pattern) }
+      pattern = pattern.downcase
+      failed_patterns << pattern if queries.none? { |sql| sql.downcase.include?(pattern) }
     end
     assert_empty failed_patterns,
       "Query pattern(s) #{failed_patterns.map(&:inspect).join(', ')} not found.#{queries.empty? ? '' : "\nQueries:\n#{queries.join("\n")}"}"
@@ -85,7 +86,8 @@ module MinitestHelpers
 
     failed_patterns = []
     patterns_to_match.each do |pattern|
-      failed_patterns << pattern if queries.any? { |sql| sql.include?(pattern) }
+      pattern = pattern.downcase
+      failed_patterns << pattern if queries.any? { |sql| sql.downcase.include?(pattern) }
     end
     assert_empty failed_patterns,
       "Query pattern(s) #{failed_patterns.map(&:inspect).join(', ')} found.#{queries.empty? ? '' : "\nQueries:\n#{queries.join("\n")}"}"